feat: :art: move stuff to module-debug and show it with a boolean debugMode
